Move UniformFloat to a new submodule
